# Performance optimization and timing/log instrumentation