- Added caching for routes lookup and network device creations
- Optimized transfer carbon cost calculation: group devices by ISOs and query carbon data once per ISO.
- Added timing instrumentation on major time consuming functions and # of loop iterations based on input list.
- Reduced per-region log message logging level.

Time improvement on daily query with all AWS + GCP regions (48 total)
| Comparison | No optimize | Optimize timing |
|---|---|---|
| Before | 13.5s | 38.7s |
| After (first run per region) | 10.3s | 36.8s |
| Before (subsequent run per region) | 4s | 23s |